(structural) splitting up the auxiliary module, oscacq -> oscilloscope, traceio -> fileio
