--- docs/contents/auxiliary.rst
@@ -1,6 +1,0 @@
-Auxiliary module :mod:`~keyoscacquire.auxiliary`
-*************************************************
-
-.. automodule:: keyoscacquire.auxiliary
-  :members:
-  :private-members:
