+++ docs/contents/dataprocessing.rst
@@ -3,18 +3,18 @@
 Data processing, file saving & loading
 **************************************
 
-.. py:currentmodule:: keyoscacquire.oscacq
-
-The :mod:`keyoscacquire.oscacq` module contains a function for processing
-the raw data captured with :class:`Oscilloscope`, and :mod:`keyoscacquire.traceio`
+The :mod:`keyoscacquire.dataprocessing` module contains a function for processing
+the raw data captured with :class:`Oscilloscope`, and :mod:`keyoscacquire.fileio`
 for saving the processed data to files and plots.
 
-Data processing
----------------
+Data processing (:mod:`keyoscacquire.dataprocessing`)
+-----------------------------------------------------
+
+.. py:currentmodule:: keyoscacquire.dataprocessing
 
 The output from the :func:`Oscilloscope.capture_and_read` function is processed
 by :func:`process_data`, a wrapper function that sends the data to the
-respective binary or ascii processing function.
+respective binary or ascii processing functions.
 
 This function is kept outside the Oscilloscope class as one might want to
 post-process data after capturing it.
@@ -22,16 +22,16 @@ post-process data after capturing it.
 .. autofunction:: process_data
 
 
-File saving and loading (:mod:`keyoscacquire.traceio`)
+File saving and loading (:mod:`keyoscacquire.fileio`)
 ------------------------------------------------------
 
-The Oscilloscope class has the method :meth:`Oscilloscope.save_trace()` for
-saving the most recently captured trace to disk. This method relies on the
-``traceio`` module.
+The Oscilloscope class has the method :meth:`keyoscacquire.Oscilloscope.save_trace()`
+for saving the most recently captured trace to disk. This method relies on the
+``fileio`` module.
 
-.. automodule:: keyoscacquire.traceio
+.. automodule:: keyoscacquire.fileio
 
-.. autofunction:: keyoscacquire.traceio.save_trace
-.. autofunction:: keyoscacquire.traceio.plot_trace
-.. autofunction:: keyoscacquire.traceio.load_trace
-.. autofunction:: keyoscacquire.traceio.load_header
+.. autofunction:: keyoscacquire.fileio.save_trace
+.. autofunction:: keyoscacquire.fileio.plot_trace
+.. autofunction:: keyoscacquire.fileio.load_trace
+.. autofunction:: keyoscacquire.fileio.load_header
