+++ docs/contents/overview.rst
@@ -2,14 +2,14 @@
 Overview and getting started
 ****************************
 
-The code is structured as a module :mod:`keyoscacquire.oscacq` containing the
+The code is structured as a module :mod:`keyoscacquire.oscilloscope` containing the
 engine doing the `PyVISA <https://pyvisa.readthedocs.io/en/latest/>`_
-interfacing in a class :class:`~keyoscacquire.oscacq.Oscilloscope`, and
+interfacing in a class :class:`~keyoscacquire.oscilloscope.Oscilloscope`, and
 support functions for data processing. Programmes are located
 in :mod:`keyoscacquire.programmes`, and the same programmes can be run
 directly from the command line as they are installed in the Python path,
 see :ref:`cli-programmes`. Default options are found in :mod:`keyoscacquire.config`,
-and the :mod:`keyoscacquire.traceio` provides functions for plotting, saving,
+and the :mod:`keyoscacquire.fileio` provides functions for plotting, saving,
 and loading traces from disk.
