+++ docs/contents/visa_utils.rst
@@ -1,0 +1,6 @@
+Visa-related support module :mod:`~keyoscacquire.visa_utils`
+************************************************************
+
+.. automodule:: keyoscacquire.visa_utils
+  :members:
+  :private-members:
